% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHDI.HDIAnalysis.R
\name{plot.HDI}
\alias{plot.HDI}
\alias{plotHDI.HDIAnalyze}
\title{Plot Method for HDIAnalyze Class}
\usage{
plotHDI.HDIAnalyze(x, ...)
}
\arguments{
\item{x}{An HDIAnalyze object.}

\item{...}{Additional arguments passed to plot.}
}
\description{
This function plots the indicators of HDIAnalyze object with values less than 100 (for better visualization).
}
\examples{
hdi_data <- read_hdi(hdro_indicators_irl.csv)
plotHDI.HDIAnalyze(hdi_data)

}
