% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print1.HDIAnalysis.R
\name{print1.HDIAnalyze}
\alias{print1.HDIAnalyze}
\title{Print method for HDIAnalyze class}
\usage{
print1.HDIAnalyze(x, ...)
}
\arguments{
\item{x}{A HDIAnalyze object.}

\item{...}{Other arguments.}
}
\description{
This function prints the first 5 observations of HDIAnalyze object. Output shows the columns of dataframe and their first five values.
}
\examples{
hdi_data <- read_hdi(hdro_indicators_ind.csv)
print1.HDIAnalyze(hdi_data)

}
