% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print2.HDIAnalysis.R
\name{print2.HDIAnalyze}
\alias{print2.HDIAnalyze}
\title{Another Print method for HDIAnalyze class}
\usage{
print2.HDIAnalyze(x, ...)
}
\arguments{
\item{x}{A HDIAnalyze object.}

\item{...}{Other arguments.}
}
\description{
This function prints the first 5 observations of HDIAnalyze object. Output shows the columns of dataframe and their first five values in a well defined tabular format.
}
\examples{
hdi_data <- read_hdi(hdro_indicators_ind.csv)
print2.HDIAnalyze(hdi_data)

}
