% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDIAnalysis.R
\name{read_hdi}
\alias{read_hdi}
\title{Read the Human Development Indicator Data}
\usage{
read_hdi(file_name)
}
\arguments{
\item{file_name}{The path to hdi.csv file}
}
\value{
An object of class hdi_analyze.
}
\description{
This function reads the Human Development Indicator (HDI) CSV file and returns a cleaned data frame.
}
\examples{
read_hdi(hdro_indicators_ind.csv)
}
