% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryHDI.HDIAnalysis.R
\name{summaryHDI}
\alias{summaryHDI}
\alias{summaryHDI.HDIAnalyze}
\title{Summary Method for HDIAnalyze Class}
\usage{
summaryHDI.HDIAnalyze(x, ...)
}
\arguments{
\item{x}{An HDIAnalyze object.}

\item{...}{Additional parameters which can be passed to summary.}
}
\description{
This function provides the summary statistics (mean, standard deviation, minimum and maximum value) of the HDIAnalyze object
sorted by the indicator_id of the object.
}
\examples{
hdi_data <- read_hdi(hdro_indicators_ind.csv)
summaryHDI.HDIAnalysis(hdi_data)

}
